Agregado ViewPager y elemento de Dias - Segundo ViewModel

+++ app/src/main/java/pe/paku/weatherapp/data/remote/dto/HourlyWeatherDTO.kt
@@ -9,11 +9,11 @@ data class HourlyWeatherDTO(
     @SerializedName("humidity")
     val humidity: Double,
     @SerializedName("rainChance")
-    val rainChance: Int,
+    val rainChance: Double,
     @SerializedName("temperature")
     val temperature: Int,
     @SerializedName("weatherType")
     val weatherType: String,
     @SerializedName("windSpeed")
-    val windSpeed: Int
+    val windSpeed: Double
 )
